requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
mongoengine==0.27.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
